+++ addOns/fuzz/src/main/java/org/zaproxy/zap/extension/fuzz/payloads/ui/impl/NumberPayloadGeneratorAdapterUIHandler.java
@@ -52,6 +52,7 @@ public class NumberPayloadGeneratorAdapterUIHandler
             Constant.messages.getString("fuzz.payloads.generator.numbers.name");
     private static final String PAYLOAD_GENERATOR_DESC =
             Constant.messages.getString("fuzz.payloads.generator.numbers.description");
+    private static final int DEFAULT_STEP = 1;
 
     @Override
     public String getName() {
@@ -166,7 +167,7 @@ public class NumberPayloadGeneratorAdapterUIHandler
 
             fromField = new ZapNumberSpinner(Integer.MIN_VALUE, 0, Integer.MAX_VALUE);
             toField = new ZapNumberSpinner(Integer.MIN_VALUE, 0, Integer.MAX_VALUE);
-            stepField = new ZapNumberSpinner(Integer.MIN_VALUE, 1, Integer.MAX_VALUE);
+            stepField = new ZapNumberSpinner(Integer.MIN_VALUE, DEFAULT_STEP, Integer.MAX_VALUE);
 
             JLabel fromLabel = new JLabel(PAYLOADS_FROM_LABEL);
             fromLabel.setLabelFor(fromField);
@@ -267,7 +268,7 @@ public class NumberPayloadGeneratorAdapterUIHandler
             getPayloadsPreviewTextArea().setText("");
             fromField.setValue(0);
             toField.setValue(0);
-            stepField.setValue(0);
+            stepField.setValue(DEFAULT_STEP);
         }
 
         @Override