+++ addOns/fuzz/CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
+### Fixed
+- The Numberzz payload generator should not set the increment to zero when creating subsequent payloads.
 
+### Changed
+- Maintenance changes.
 
 ## [13.2.0] - 2021-06-01
 ### Changed